Add files via upload
In the presented project, a Text to Music Generator using Python, leveraging deep learning techniques and streamlit for interactive web-based deployment. The aim is to allow users to input text descriptions or lyrics and generate corresponding musical compositions in various genres. The system utilizes pre-trained neural network models specifically designed for music generation, such as MusicGen, provided by the Audiocraft library. The generated music is then saved as audio files in WAV format, allowing users to listen to and download the compositions. The system architecture involves a backend powered by PyTorch and torchaudio for deep learning tasks, while the frontend is built using the Streamlit library for creating intuitive user interfaces. Users are presented with a simple web interface where they can input their desired text description and select the duration of the generated music. Upon clicking the "Generate Music" button, the system processes the input text, generates music tensors using the pre-trained model, and saves the resulting audio files locally. Key features of the system include a sidebar displaying available song genres, a text area for inputting descriptions or lyrics, and a slider for selecting the duration of the generated music. Users can choose from a variety of genres such as Pop, Rock, Jazz, Classical, Hip- hop, Electronic, and Country. After generating the music, users are presented with the option to listen to the compositions directly on the web interface and download the audio files for offline use. The system's workflow involves loading the pre-trained MusicGen model, setting generation parameters such as sampling techniques and duration, generating music tensors based on the input description, and saving the resulting audio files. The code is structured into modular functions for better organization and maintainability, with caching mechanisms implemented to optimize performance.